+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactor of Editor code
 - Removed "Buffers" from Scripts
 - Refactor of IsApproximately
+- Updated RoadUtility to use Unity 2019.2 API
 
 
 ## [Version: 1.9_FH] - 2019.04.22